packaging: Generate uber jars for -core and -apache
Relocating embedded dependencies inside com.algoliasearch, filtering
out unused resources.

Note that slf4j-api is NOT embedded in the uber jars, on purpose: if
slf4j-api is embedded here, it would need to be relocated (*), yet if
gets relocated, implementations cannot be injected anymore (**),
making slf4j irrelevant.

(*) Embedding jars in an uber-library _require_ their relocation: a
library is meant to be used along with other jars, which may depend on
other versions of the jars embedded here. Relocating these mostly get
rid of conflicts; not relocating them, would on the contrary make
conflicts practically impossible to solve.

(**) the injection used by slf4j is namespace based; if the api gets
relocated, implementation would have to be relocated too.

From a "pom.xml point of view", a pluginManagement section is set in
the root pom.xml to configure the shading - but not apply it. This
configuration in then leveraged in dedicated modules
(algolia-core-uber and algolia-apache-uber) coming with their own
pom.xml files (and possibly adjusting - augmenting - the relocations,
in the apache flavor).