feat(apache): improve performances by changing maxConnPerRoute from 2 to 100
After some investigations in #678, we discovered that our default value
of `maxConnPerRoute` was set to 2. This led to poor performances
compared to our previous v1 major version. After some benchmarking, we
noticed that this default value, while being conservative, is probably
to low for most use-cases.

Note that we can safely change this value since it can now be overriden
by passing a custom `HttpAsyncClientBuilder` during the `ApacheHttpRequester`
instanciation (see our README for more information).