chore: fix fmt-maven-plugin configuration
Previously, the fmt-maven-plugin configuration was located under the
<build> configuration of our default profile. This lead to our
configured version of our formater to be completely ignored. Later on,
an error was discovered since a very recent version of our formater was
fetched instead, whose one of its transitive dependencies
(google-java-format) was upgraded (1.8) to a version (1.8) which drop
the Java 8 support.

This commit correctly sets up the fmt-maven-plugin configuration within
the <build><pluginManagement><plugins> tree instead. It also updates the
plugin from 2.8 to 2.9 and adds a comment to the pom.xml indicated *not
to* upgrade to 2.10, this version being the one whose transitive
dependencies drop Java 8 support.