feat(build): make Apache Uber JAR OSGi-compliant
Following the requests made by some recent users, this commit changes
the `pom.xml` configuration of our `algoliasearch-apache-uber` module in
order to make it installable as a valid OSGi bundle.

To ensure this works, we have created a new `example-osgi` directory.
This project is a "Hello World"-like OSGi bundle application which
depends on the `algoliasearch-apache-uber` module at runtime. It
contains both the code and a detailled `README.md` on how to install,
build and deploy such application from scratch, using the Apache Karaf
OSGi-application runner.