chore(ci): make CircleCI perform the `mvn deploy` step upon tagged commit on master branch
This commit adds an extra step to the CircleCI configuration to release
the project to Bintray if and only if the CI is passing on a semver
tagged commit from master and all tests are passing.

In essence, it prevents the releaser to locally configure Maven and
fetch credentials and let CircleCI perform the `mvn deploy` on its own.

You can find extra configuration information from the official CircleCI
blog at this address:
https://circleci.com/blog/optimizing-maven-builds-on-circleci/